better visual effect: moving bound and flying out.

+++ game/config.h
@@ -1,10 +1,11 @@
 #ifndef CONFIG_H
 #define CONFIG_H
 
+#include "qglobal.h"
 const int WIDTH = 800;
 const int HEIGHT = 1000;
-const float FPS = 60;
-const float BULLET_SPEED = 10;
-const float PLAYER_MOVING_SPEED = 10;
+const qreal FPS = 60;
+const qreal BULLET_SPEED = 10;
+const qreal PLAYER_MOVING_SPEED = 10;
 
 #endif // CONFIG_H
